+++ a2a-grpc/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.5](https://github.com/a2aproject/a2a-rs/compare/a2a-grpc-v0.3.4...a2a-grpc-v0.3.5) - 2026-08-27
+
+### Other
+
+- updated the following local packages: a2a-client-lf, a2a-server-lf
+
+## [0.3.4](https://github.com/a2aproject/a2a-rs/compare/a2a-grpc-v0.3.3...a2a-grpc-v0.3.4) - 2026-08-26
+
+### Other
+
+- updated the following local packages: a2a-server-lf
+
+## [0.3.3](https://github.com/a2aproject/a2a-rs/compare/a2a-grpc-v0.3.2...a2a-grpc-v0.3.3) - 2026-08-25
+
+### Other
+
+- updated the following local packages: a2a-client-lf
+
 ## [0.3.2](https://github.com/a2aproject/a2a-rs/compare/a2a-grpc-v0.3.1...a2a-grpc-v0.3.2) - 2026-07-27
 
 ### Fixed